+++ core/src/server/Config.h
@@ -28,6 +28,14 @@
 namespace milvus {
 namespace server {
 
+#define CONFIG_CHECK(func) \
+    do {                   \
+        Status s = func;   \
+        if (!s.ok()) {     \
+            return s;      \
+        }                  \
+    } while (false)
+
 static const char* CONFIG_NODE_DELIMITER = ".";
 static const char* CONFIG_VERSION = "version";
 
@@ -58,6 +66,21 @@ static const char* CONFIG_DB_INSERT_BUFFER_SIZE = "insert_buffer_size";
 static const char* CONFIG_DB_INSERT_BUFFER_SIZE_DEFAULT = "1";
 static const char* CONFIG_DB_PRELOAD_TABLE = "preload_table";
 static const char* CONFIG_DB_PRELOAD_TABLE_DEFAULT = "";
+
+/* storage config */
+static const char* CONFIG_STORAGE = "storage_config";
+static const char* CONFIG_STORAGE_MINIO_ENABLE = "minio_enable";
+static const char* CONFIG_STORAGE_MINIO_ENABLE_DEFAULT = "false";
+static const char* CONFIG_STORAGE_MINIO_ADDRESS = "minio_address";
+static const char* CONFIG_STORAGE_MINIO_ADDRESS_DEFAULT = "127.0.0.1";
+static const char* CONFIG_STORAGE_MINIO_PORT = "minio_port";
+static const char* CONFIG_STORAGE_MINIO_PORT_DEFAULT = "9000";
+static const char* CONFIG_STORAGE_MINIO_ACCESS_KEY = "minio_access_key";
+static const char* CONFIG_STORAGE_MINIO_ACCESS_KEY_DEFAULT = "minioadmin";
+static const char* CONFIG_STORAGE_MINIO_SECRET_KEY = "minio_secret_key";
+static const char* CONFIG_STORAGE_MINIO_SECRET_KEY_DEFAULT = "minioadmin";
+static const char* CONFIG_STORAGE_MINIO_BUCKET = "minio_bucket";
+static const char* CONFIG_STORAGE_MINIO_BUCKET_DEFAULT = "milvus-bucket";
 
 /* cache config */
 static const char* CONFIG_CACHE = "cache_config";
@@ -169,6 +192,20 @@ class Config {
     Status
     CheckDBConfigInsertBufferSize(const std::string& value);
 
+    /* storage config */
+    Status
+    CheckStorageConfigMinioEnable(const std::string& value);
+    Status
+    CheckStorageConfigMinioAddress(const std::string& value);
+    Status
+    CheckStorageConfigMinioPort(const std::string& value);
+    Status
+    CheckStorageConfigMinioAccessKey(const std::string& value);
+    Status
+    CheckStorageConfigMinioSecretKey(const std::string& value);
+    Status
+    CheckStorageConfigMinioBucket(const std::string& value);
+
     /* metric config */
     Status
     CheckMetricConfigEnableMonitor(const std::string& value);
@@ -243,6 +280,20 @@ class Config {
     Status
     GetDBConfigPreloadTable(std::string& value);
 
+    /* storage config */
+    Status
+    GetStorageConfigMinioEnable(bool& value);
+    Status
+    GetStorageConfigMinioAddress(std::string& value);
+    Status
+    GetStorageConfigMinioPort(std::string& value);
+    Status
+    GetStorageConfigMinioAccessKey(std::string& value);
+    Status
+    GetStorageConfigMinioSecretKey(std::string& value);
+    Status
+    GetStorageConfigMinioBucket(std::string& value);
+
     /* metric config */
     Status
     GetMetricConfigEnableMonitor(bool& value);
@@ -310,6 +361,20 @@ class Config {
     SetDBConfigArchiveDaysThreshold(const std::string& value);
     Status
     SetDBConfigInsertBufferSize(const std::string& value);
+
+    /* storage config */
+    Status
+    SetStorageConfigMinioEnable(const std::string& value);
+    Status
+    SetStorageConfigMinioAddress(const std::string& value);
+    Status
+    SetStorageConfigMinioPort(const std::string& value);
+    Status
+    SetStorageConfigMinioAccessKey(const std::string& value);
+    Status
+    SetStorageConfigMinioSecretKey(const std::string& value);
+    Status
+    SetStorageConfigMinioBucket(const std::string& value);
 
     /* metric config */
     Status